display_main에도 url 파라미터 추가
